services/api/pyproject.toml
[project]
name = "news_api"
version = "0.1.0"
requires-python = ">=3.12"
description = "API + Auth service — FastAPI on Lambda + Clerk JWT validation"
dependencies = [
    "fastapi>=0.115",
    "mangum>=0.19",
    "pyjwt[crypto]>=2.9",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.5",
    "loguru>=0.7.3",
    "typer>=0.24.2",
    "uvicorn>=0.30",
    "httpx>=0.28.1",
    "boto3>=1.42.95",
    "news_schemas",
    "news_config",
    "news_observability",
    "news_db",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/news_api"]
